refactor(encrypted-archive): reduce lines in `src/key-derivation-function/argon2.ts` file
Code Climate reported:
+ File `argon2.ts` has 281 lines of code (exceeds 250 allowed). Consider refactoring.
